+++ packages/services/api/src/modules/target/resolvers/Project.ts
@@ -1,7 +1,7 @@
 import { TargetManager } from '../providers/target-manager';
 import type { ProjectResolvers } from './../../../__generated__/types.next';
 
-export const Project: Pick<ProjectResolvers, 'targets'> = {
+export const Project: Pick<ProjectResolvers, 'targets' | '__isTypeOf'> = {
   targets: (project, _, { injector }) => {
     return injector.get(TargetManager).getTargets({
       project: project.id,